refactor: 중복 선택 버튼 리팩토링
